0.2.2

+++ src/index.js
@@ -25,11 +25,13 @@ const lock = new K8SLock({
   waitUntilLock: true,
   createLeaseIfNotExist: true,
   leaseDurationInSeconds: 15,
-  refreshLockInterval: 3000,
+  refreshLockInterval: 500,
   lockTryInterval: 3000
 });
 
+// set debug mode, false by default
 const debugMode = process.env.DEBUG_MODE || 'false';
+// global variables
 var applyingScheduled = false;
 var settingStatus = new Map();
 var lastClusterResourceVersion = '';
@@ -39,16 +41,25 @@ var collectionWatch = '';
 var clusterWatchStart = true;
 var collectionWatchStart = true;
 var lockId = '';
+var lockInfo = '';
+// load KubeConfig
 const kc = new k8s.KubeConfig();
 kc.loadFromDefault();
 
+// initialize various K8S APIs
 const k8sCoreApi = kc.makeApiClient(k8s.CoreV1Api);
 const k8sCustomApi = kc.makeApiClient(k8s.CustomObjectsApi);
 const k8sPolicyApi = kc.makeApiClient(k8s.PolicyV1Api);
 const k8sAppsApi = kc.makeApiClient(k8s.AppsV1Api);
 const watch = new k8s.Watch(kc);
 
+// react on QdrantClusters events
 const onEventCluster = async (phase, apiObj) => {
+  // leader status was lost
+  if (!lockInfo.isLocking) {
+    log('Leader status was lost, restarting...');
+    process.exit(1);
+  }
   // ignore MODIFIED on status changes
   if (settingStatus.has(apiObj.metadata.name)) {
     return;
@@ -57,43 +68,56 @@ const onEventCluster = async (phase, apiObj) => {
   if (lastClusterResourceVersion == apiObj.metadata.resourceVersion) {
     return;
   }
+  // update ResourceVersion for the last caught cluster
   lastClusterResourceVersion = apiObj.metadata.resourceVersion;
   log(`Received event in phase ${phase}.`);
-
+  // start applyting cluster changes
   if (['ADDED', 'MODIFIED'].includes(phase)) {
     try {
       scheduleApplying(apiObj);
     } catch (err) {
       log(err);
     }
+    // will be cleaned automatically by ownerReferences
   } else if (phase == 'DELETED') {
     log(`${apiObj.kind} "${apiObj.metadata.name}" was deleted!`);
   }
 };
 
+// react on QdrantCollections events
 const onEventCollection = async (phase, apiObj) => {
+  // leader status was lost
+  if (!lockInfo.isLocking) {
+    log('Leader status was lost, restarting...');
+    process.exit(1);
+  }
   // ignore duplicated event on watch reconnections
   if (lastCollectionResourceVersion == apiObj.metadata.resourceVersion) {
     return;
   }
+  // update ResourceVersion for the last caught collection
   lastCollectionResourceVersion = apiObj.metadata.resourceVersion;
   log(`Received event in phase ${phase}.`);
-
+  // wait for collection creation
   if (phase == 'ADDED') {
     await createCollection(apiObj, k8sCustomApi, k8sCoreApi);
+    // wait for collection update
   } else if (phase == 'MODIFIED') {
     await updateCollection(apiObj, k8sCustomApi, k8sCoreApi);
+    // wait for collection delete
   } else if (phase == 'DELETED') {
     await deleteCollection(apiObj, k8sCustomApi, k8sCoreApi);
   }
 };
 
+// QdrantClusters watch has stopped unexpectedly, restart
 const onDoneCluster = (err) => {
   log(`Connection to QdrantClusters closed, reconnecting...`);
   clusterWatchStart = true;
   watchResource();
 };
 
+// QdrantCollections watch has stopped unexpectedly, restart
 const onDoneCollection = (err) => {
   log(`Connection to QdrantCollections closed, reconnecting...`);
   collectionWatchStart = true;
@@ -101,7 +125,7 @@ const onDoneCollection = (err) => {
 };
 
 const watchResource = async () => {
-  //restart required watchers
+  //restart required watches
   var watchList = [];
   if (clusterWatchStart) {
     watchList.push(
@@ -127,14 +151,17 @@ const watchResource = async () => {
     log('Watching QdrantCollections API.');
     collectionWatchStart = false;
   }
-
+  // return the first caught event from any watch
   return Promise.any(watchList);
 };
 
+// set the customresource status, clusters only at the moment
 const setStatus = async (apiObj, k8sCustomApi, status) => {
   const name = apiObj.metadata.name;
   const namespace = apiObj.metadata.namespace;
+  // add cluster name to the map of currently updating resources
   settingStatus.set(name, 'update');
+  // get current status
   const readObj = await k8sCustomApi.getNamespacedCustomObjectStatus(
     'qdrant.operator',
     'v1alpha1',
@@ -143,6 +170,7 @@ const setStatus = async (apiObj, k8sCustomApi, status) => {
     name
   );
   const resCurrent = readObj.body;
+  // prepare new payload
   const newStatus = {
     apiVersion: apiObj.apiVersion,
     kind: apiObj.kind,
@@ -155,6 +183,7 @@ const setStatus = async (apiObj, k8sCustomApi, status) => {
     }
   };
   try {
+    // set new status
     const res = await k8sCustomApi.replaceNamespacedCustomObjectStatus(
       'qdrant.operator',
       'v1alpha1',
@@ -167,9 +196,11 @@ const setStatus = async (apiObj, k8sCustomApi, status) => {
   } catch (err) {
     log(err);
   }
+  // job is done, remove this resource from the map
   settingStatus.delete(name);
 };
 
+// update the version of last caught cluster
 const updateResourceVersion = async (apiObj, k8sCustomApi) => {
   const name = apiObj.metadata.name;
   const namespace = apiObj.metadata.namespace;
@@ -184,24 +215,31 @@ const updateResourceVersion = async (apiObj, k8sCustomApi) => {
   lastClusterResourceVersion = resCurrent.metadata.resourceVersion;
 };
 
+// check the cluster readiness
 const waitForClusterReadiness = (apiObj, k8sAppsApi, k8sCustomApi) => {
+  // start background periodic job
   let interval = setInterval(
     async function (apiObj, k8sAppsApi, k8sCustomApi) {
       const name = apiObj.metadata.name;
       const namespace = apiObj.metadata.namespace;
       try {
+        // get qdrant statefulset
         const res = await k8sAppsApi.readNamespacedStatefulSet(
           `${name}`,
           `${namespace}`
         );
         const stset = res.body;
+        // wait until available and updated replicas >= desired replicas count
         if (
           stset.status.availableReplicas >= stset.spec.replicas &&
           stset.status.updatedReplicas >= stset.spec.replicas
         ) {
           log(`Cluster "${name}" is ready!`);
+          // set resource status to Running
           await setStatus(apiObj, k8sCustomApi, 'Running');
+          // memorize tis resourceversion
           await updateResourceVersion(apiObj, k8sCustomApi);
+          // stop watching
           clearInterval(interval);
         } else {
           log(
@@ -210,7 +248,9 @@ const waitForClusterReadiness = (apiObj, k8sAppsApi, k8sCustomApi) => {
         }
         return;
       } catch (err) {
+        // can't read statefulset, probably it was killed
         log(`Cluster "${name}" was terminated, stop watching.`);
+        // stop watching
         clearInterval(interval);
       }
     },
@@ -228,6 +268,7 @@ const scheduleApplying = (apiObj) => {
   }
 };
 
+// create all required k8s resources
 const applyNow = async (apiObj) => {
   applyingScheduled = false;
   await setStatus(apiObj, k8sCustomApi, 'Pending');
@@ -244,20 +285,23 @@ const applyNow = async (apiObj) => {
 };
 
 const main = async () => {
-  // leader election using leases
+  // leader election using k8s leases
   log(
     `Status of "${process.env.POD_NAME}": FOLLOWER. Trying to get leader status...`
   );
-  const lockInfo = await lock.startLocking();
+  lockInfo = await lock.startLocking();
   lockId = lockInfo.lockId;
   log(`Status of "${process.env.POD_NAME}": LEADER.`);
+  // start watching events only after taking ownership of the lease
   await watchResource();
 };
 
+// log format
 export const log = (message) => {
   console.log(`${new Date().toLocaleString()}: ${message}`);
 };
 
+// print all errors
 if (debugMode == 'true') {
   log('Debug mode ON!');
   process.on('unhandledRejection', (reason, p) => {
@@ -265,6 +309,7 @@ if (debugMode == 'true') {
   });
 }
 
+// got SIGTERM - stop locking and exit
 process.on('SIGTERM', async () => {
   await lock.stopLocking(lockId);
   log('Stopping gracefully...');
